backend/action_engine/requirements.txt
functions-framework==3.*
google-cloud-secret-manager==2.*
google-cloud-storage==2.*
requests==2.*
